% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissueSpecificRegressionHelpers.R
\name{computeEta}
\alias{computeEta}
\title{Linear predictor}
\usage{
computeEta(params, X, log_tech_factors, tissue_size_factors)
}
\arguments{
\item{params}{Vector of model parameters: regression coefficients, soft-maxed proportions and inverse overdispersion}

\item{X}{design \code{matrix} (output of \code{model.matrix(mean_design, annots)})}

\item{log_tech_factors}{Log of Technical factors (i.e. log of normalization factors for ERCC-spikeins)}

\item{tissue_size_factors}{Tissue-specific size factor (normalization factors for tissue-unique genes divided by normalization factors for ERCC-spikeins)}
}
\value{
Linear predictor
}
\description{
Estimates the mean of the model (i.e. the linear predictor)
given the parameter
}
