% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissueSpecificRegressionHelpers.R
\name{computeNegativeLogLikelihoodGradient}
\alias{computeNegativeLogLikelihoodGradient}
\title{Negative log-likelihood gradients}
\usage{
computeNegativeLogLikelihoodGradient(
  params,
  y,
  X,
  log_tech_factors,
  tissue_size_factors,
  theta,
  model
)
}
\arguments{
\item{params}{Vector of model parameters: regression coefficients, soft-maxed proportions and inverse overdispersion}

\item{y}{\code{integer} counts of a gene}

\item{X}{design \code{matrix} (output of \code{model.matrix(mean_design, annots)})}

\item{log_tech_factors}{Log of Technical factors (i.e. log of normalization factors for ERCC-spikeins)}

\item{tissue_size_factors}{Tissue-specific size factor (normalization factors for tissue-unique genes divided by normalization factors for ERCC-spikeins)}

\item{theta}{Inverse of over-dispersion parameter (size parameter in \link[stats]{dnbinom})}

\item{model}{Statistical distribution: 'nb' for Negative Binomial or 'poisson' for Poisson}
}
\value{
Negative log-likelihood gradients of model parameters
}
\description{
Estimates negative log-likelihood gradients of regression
coefficients and soft-maxed proportions
}
