% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissueSpecificRegressionHelpers.R
\name{fitModel}
\alias{fitModel}
\title{Model fitting and statistical testing}
\usage{
fitModel(
  y,
  X,
  tech_factors,
  tissue_size_factors,
  max_iterations = 10,
  wald_test = TRUE
)
}
\arguments{
\item{y}{\code{integer} counts of a gene}

\item{X}{design \code{matrix} (output of \code{model.matrix(mean_design, annots)})}

\item{tech_factors}{Technical factors (i.e. normalization factors for ERCC-spikeins)}

\item{tissue_size_factors}{Tissue-specific size factor (normalization factors for tissue-unique genes divided by normalization factors for ERCC-spikeins)}

\item{max_iterations}{\code{integer} specifying maximum number of iterations for maximum-likelihood estimation}

\item{wald_test}{\code{logical} specifying is Wald test should be performed on regression coefficients}
}
\value{
Maximum likelihood estimates of negative binomial model
and p-values for regression coefficients if requested
}
\description{
Fit negative binomial model using maximum likelihood for a
single gene and run Wald test if requested
}
