% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissueSpecificRegressionHelpers.R
\name{initParams}
\alias{initParams}
\title{Initialize parameters}
\usage{
initParams(X, tissue_size_factors)
}
\arguments{
\item{X}{design \code{matrix} (output of \code{model.matrix(mean_design, annots)})}

\item{tissue_size_factors}{Tissue-specific size factor (normalization factors for tissue-unique genes divided by normalization factors for ERCC-spikeins)}
}
\value{
Random starting parameters
}
\description{
Random starting regression coefficients and
soft-maxed proportions used for model fitting
}
