% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalizationFactor}
\alias{normalizationFactor}
\title{Standard normalization factors}
\usage{
normalizationFactor(counts, ...)
}
\arguments{
\item{counts}{RNA-seq biological count \code{matrix} (genes as rows, samples as columns)}

\item{...}{Furthur arguments passed to \link[scran]{computeSumFactors}}
}
\value{
\code{numeric} \code{vector} of normalization factors
}
\description{
Estimate standard normalization factors using \link[scran]{computeSumFactors}
}
