% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalizeCounts}
\alias{normalizeCounts}
\title{Standard normalization}
\usage{
normalizeCounts(counts, nf)
}
\arguments{
\item{counts}{RNA-seq biological count \code{matrix} (genes as rows, samples as columns)}

\item{nf}{Standard normalization factors (output of \link[isotela]{normalizationFactor})}
}
\value{
Normalized read counts
}
\description{
Normalize counts using standard normalization
}
