% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalizeCountsTissueSpecific}
\alias{normalizeCountsTissueSpecific}
\title{Tissue-specific normalization}
\usage{
normalizeCountsTissueSpecific(counts, nf_list, tissues)
}
\arguments{
\item{counts}{RNA-seq biological count \code{matrix} (genes as rows, samples as columns)}

\item{nf_list}{Standard normalization factors (output of \link[isotela]{normalizationFactorTissueSpecific})}

\item{tissues}{\code{vector} with tissues as values and genes as names (have to match \code{rownames(counts)})}
}
\value{
Tissue-normalized read counts of tissue-unique genes
}
\description{
Normalize counts using tissue-specific normalization
}
